20260320_face_balance/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.3.0
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
numpy==1.26.4
scipy==1.13.0
face_recognition==1.3.0
boto3==1.34.131
python-multipart==0.0.9
pillow==10.3.0
PyYAML==6.0.1
